NAS-141134: Extract TN_TABLE_PAGER_LABELS provider
